requirements.txt
requests
beautifulsoup4
pandas
apache-airflow
